Reorganize types dir a bit
