--- packages/core/src/types/common/index.ts
@@ -1,4 +1,0 @@
-export * from './db'
-export * from './eth'
-export * from './net'
-export * from './utils'
