+++ packages/core/src/types/index.ts
@@ -1,7 +1,8 @@
 export * from './ovm'
-export * from './common'
+export * from './miscellaneous'
 export * from './keystore'
 export * from './block-production.types'
 export * from './transport.interface'
 export * from './type-guards'
 export * from './db'
+export * from './serialization'
