+++ packages/core/src/types/miscellaneous/crypto.interface.ts
