+++ packages/core/src/types/miscellaneous/eth/event-watcher.interface.ts
