+++ packages/core/src/types/miscellaneous/eth/keystore.interface.ts
