+++ packages/core/src/types/serialization/index.ts
@@ -1,3 +1,2 @@
-export * from './crypto.interface'
 export * from './state.interface'
 export * from './type.interface'
