+++ packages/core/src/types/serialization/state.interface.ts
