+++ packages/core/src/types/serialization/type.interface.ts
